% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputChecks.R
\name{.requireConstant}
\alias{.requireConstant}
\title{Internal function to enforce input is a single number}
\usage{
.requireConstant(n, name)
}
\description{
Internal function to enforce input is a single number
}
\keyword{internal}
