% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputChecks.R
\name{.requireWhole}
\alias{.requireWhole}
\title{Internal function for checking that a number is a whole number.}
\usage{
.requireWhole(n, name)
}
\description{
Internal function for checking that a number is a whole number.
}
