% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimBlobs.R
\name{SimBlobs}
\alias{SimBlobs}
\title{Blobs}
\usage{
SimBlobs(N, nFeatures = 3, centers = NULL, centerBox = c(-10, 10), seed = NULL)
}
\arguments{
\item{N}{If an integer the number of data points per blob, if a vector the number
of data points for each blob.}

\item{nFeatures}{The number of features for each blob (default = 3).}

\item{centers}{The centers of the blobs (default = NULL).}

\item{centerBox}{The bounding box for centers when generated randomly (default = c(-10, 10)).}

\item{seed}{seed for RNG (default = NULL).}
}
\value{
list(x = x, y = y, centers = centers)
}
\description{
Simulate from isometric Gaussian blobs.
}
\examples{
N = c(2,2,2)
centers = centers = matrix(c(1,10,35, 67,6,7), ncol = 2)
SimBlobs(N, centers = centers)
}
