% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFriedman2.R
\name{SimFriedman2}
\alias{SimFriedman2}
\title{Friedman 2}
\usage{
SimFriedman2(N, sigma, seed = NULL)
}
\arguments{
\item{N}{Number of data points to simulate. Positive integer.}

\item{sigma}{Noise level for Gaussian noise. Must be postive.}

\item{seed}{seed for RNG (default = NULL).}
}
\value{
list(x = x, y = y)
}
\description{
Simulate from the 'Friedman #2' regression problem.
}
\examples{
N = 100
sigma = 0.1
SimFriedman2(N, sigma)
}
