% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPolyChangePoint.R
\name{SimPolyChangePoint}
\alias{SimPolyChangePoint}
\title{Simulate polynomial change point model.}
\usage{
SimPolyChangePoint(
  N,
  sigma = NULL,
  xRange = NULL,
  x0 = NULL,
  m1 = NULL,
  m2 = NULL
)
}
\arguments{
\item{N}{number of data points to simulate.}

\item{sigma}{noise level, default = .1.}

\item{xRange}{minimum and maximum x values, default = c(0,1).}

\item{x0}{change point value, default = 0.5.}

\item{m1}{coefficients for first polynomial model, default = c(1,2).}

\item{m2}{coefficients for second polynomial model, default = c(0,3).}
}
\value{
list(data = matrix(x,y,z), t = t)
}
\description{
Simulate data from a change point model made of polynomials.
}
\examples{
N = 100
noise = .1
xRange = c(0,1)
x0 = .5
m1 = c(1,2)
m2 = c(5,2)
SimPolyChangePoint(N, noise, xRange, x0, m1, m2)
}
