% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSwissRoll.R
\name{SimSwissRoll}
\alias{SimSwissRoll}
\title{Simulate a Swiss Roll in 3d}
\usage{
SimSwissRoll(N, sigma, seed = NULL)
}
\arguments{
\item{N}{number of data points to simulate.}

\item{sigma}{noise level.}

\item{seed}{seed for RNG (default = NULL).}
}
\value{
list(data = matrix(x,y,z), t = t)
}
\description{
Simulate data from a Swiss Roll.
}
\examples{
N = 100
noise = .1
SimSwissRoll(N, noise)
}
